Tests/LibWeb/Text/expected/fieldset-client-dimensions.txt

with legend: clientTop=2 clientHeight=44
no legend: clientTop=2 clientHeight=28
with legend: paddingTop=5px
no legend: paddingTop=5px
